# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MMP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/postback-rules": {
      "get": {
        "tags": [
          "postbacks"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_v1_postback_rules_get",
        "parameters": [
          {
            "name": "app_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostbackRuleOut"
                  },
                  "title": "Response List Rules V1 Postback Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "postbacks"
        ],
        "summary": "Create Rule",
        "operationId": "create_rule_v1_postback_rules_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostbackRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostbackRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/postback-rules/{rule_id}": {
      "patch": {
        "tags": [
          "postbacks"
        ],
        "summary": "Update Rule",
        "operationId": "update_rule_v1_postback_rules__rule_id__patch",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostbackRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostbackRuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "postbacks"
        ],
        "summary": "Disable Rule",
        "description": "Disable, not delete.\n\nDelivery rows reference this rule, and a deleted rule would leave an\nadvertiser unable to explain conversions their network already received.",
        "operationId": "disable_rule_v1_postback_rules__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/postback-rules/{rule_id}/preview": {
      "get": {
        "tags": [
          "postbacks"
        ],
        "summary": "Preview Rule",
        "description": "Render the templates against sample values.\n\nLets someone see the URL their partner will actually receive before any real\nconversion depends on it \u2014 which is the difference between finding a\nmisplaced parameter now and finding it in a reconciliation meeting.",
        "operationId": "preview_rule_v1_postback_rules__rule_id__preview_get",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostbackRuleCreate": {
        "properties": {
          "app_id": {
            "type": "string",
            "format": "uuid",
            "title": "App Id"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "trigger_event": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Trigger Event"
          },
          "url_template": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url Template"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST"
            ],
            "title": "Method",
            "default": "GET"
          },
          "body_template": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Template"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "success_status_codes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Success Status Codes"
          },
          "requires_attribution": {
            "type": "boolean",
            "title": "Requires Attribution",
            "default": true
          },
          "is_sandbox": {
            "type": "boolean",
            "title": "Is Sandbox",
            "default": false
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          }
        },
        "type": "object",
        "required": [
          "app_id",
          "name",
          "trigger_event",
          "url_template"
        ],
        "title": "PostbackRuleCreate"
      },
      "PostbackRuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "app_id": {
            "type": "string",
            "format": "uuid",
            "title": "App Id"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          },
          "provider_integration_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Integration Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "trigger_event": {
            "type": "string",
            "title": "Trigger Event"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "url_template": {
            "type": "string",
            "title": "Url Template"
          },
          "body_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Template"
          },
          "success_status_codes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Success Status Codes"
          },
          "requires_attribution": {
            "type": "boolean",
            "title": "Requires Attribution"
          },
          "is_sandbox": {
            "type": "boolean",
            "title": "Is Sandbox"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "header_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Header Names"
          }
        },
        "type": "object",
        "required": [
          "id",
          "app_id",
          "campaign_id",
          "provider_integration_id",
          "name",
          "trigger_event",
          "method",
          "url_template",
          "body_template",
          "success_status_codes",
          "requires_attribution",
          "is_sandbox",
          "enabled",
          "created_at"
        ],
        "title": "PostbackRuleOut"
      },
...
      "TemplatePreview": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "variables_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variables Used"
          },
          "available_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Variables"
          }
        },
        "type": "object",
        "required": [
          "url",
          "body",
          "variables_used",
          "available_variables"
        ],
        "title": "TemplatePreview"
      },
...
    }
  }
}